64bit compilation fix
Fixes mismatching feedback key for railcom returns when running on a 64-bit processor.
Makes the dcc_packet and dcc_feedback structures contain a uintptr_t for the feedback key. This will change on a 64-bit processor.

+++ src/dcc/railcom.h
@@ -55,7 +55,7 @@ typedef struct dcc_feedback
     /// Opaque identifier that allows linking outgoing dcc::Packet sent to the
     /// DCC waveform generator to the incoming dcc::Feedback structure read
     /// back from the railcom driver.
-    uint32_t feedbackKey;
+    uintptr_t feedbackKey;
 } DCCFeedback;
 
 #ifdef __cplusplus
